Fix more 32bit -> 64bit occurences.

+++ src/search/seed_array/seed_array.h
@@ -141,7 +141,7 @@ struct SeedArray
 private:
 
 	Entry *data_;
-	std::vector<int64_t> begin_;
+	std::vector<size_t> begin_;
 	std::vector<std::vector<Entry>> entries_;
 	Search::SeedStats stats_;
